move Transit.receive into cmd/Main
