+++ src/Transit.hs
@@ -20,12 +20,17 @@
 --
 module Transit
   ( FileTransfer.sendFile
-  , FileTransfer.receive
+  , FileTransfer.receiveFile
   , FileTransfer.MessageType(..)
   , Peer.sendOffer
+  , Peer.receiveOffer
   , Peer.receiveMessageAck
+  , Peer.sendMessageAck
+  , Peer.decodeTransitMsg
+  , Network.CommunicationError(..)
   )
 where
 
 import qualified Transit.Internal.FileTransfer as FileTransfer
 import qualified Transit.Internal.Peer as Peer
+import qualified Transit.Internal.Network as Network
